Fixes a bug in new project form
The change event was binded to all select elements in project.js,
causing that each change in a combo when creating new projects, was
submiting the form and hence, creating new projects. Now this select
elements has been scoped properly (to the combos in projects
search form) to avoid this issue.